custom video upload extension structure
