+++ packages/plugin-video/src/icons/PlusIcon.jsx
@@ -1,0 +1,19 @@
+/* eslint-disable max-len */
+import React from 'react';
+
+const PlusIcon = props => (
+    <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="11"
+        height="11"
+        viewBox="0 0 11 11"
+        style={{display: 'inline-block'}}
+        {...props}
+    >
+        <path fill="#236FFF" fillRule="evenodd" d="M6 5V.2a.2.2 0 0 0-.2-.2h-.6a.2.2 0 0 0-.2.2V5H.2a.2.2 0 0 0-.2.2v.6c0 .11.09.2.2.2H5v4.8c0 .11.09.2.2.2h.6a.2.2 0 0 0 .2-.2V6h4.8a.2.2 0 0 0 .2-.2v-.6a.2.2 0 0 0-.2-.2H6z" />
+    </svg>
+
+
+);
+
+export default PlusIcon;
