+++ packages/plugin-video/src/icons/index.js
@@ -1,3 +1,4 @@
 export { default as InsertPluginIcon } from './InsertPluginIcon';
 export { default as MediaReplaceIcon } from './MediaReplaceIcon';
 export { default as VideoCameraIcon } from './VideoCameraIcon';
+export { default as PlusIcon } from './PlusIcon';
