+++ packages/plugin-video/src/toolbar/createToolbar.js
@@ -1,10 +1,10 @@
 import createInlineButtons from './inline-video-buttons';
 import createInsertButtons from './insert-video-buttons';
 
-export default function createToolbar({ helpers, t }) {
+export default function createToolbar({ helpers, t, settings }) {
   return {
-    InlineButtons: createInlineButtons({ t }),
-    InsertButtons: createInsertButtons({ helpers, t }),
+    InlineButtons: createInlineButtons({ t, settings }),
+    InsertButtons: createInsertButtons({ helpers, t, settings }),
     name: 'video',
   };
 }
