+++ packages/plugin-video/src/toolbar/inline-video-buttons.jsx
@@ -1,8 +1,8 @@
-import { BUTTONS, getModalStyles } from 'wix-rich-content-common';
+import { BUTTONS, getModalStyles, decorateComponentWithProps } from 'wix-rich-content-common';
 import { Modals } from '../modals';
 import { MediaReplaceIcon } from '../icons';
 
-export default({ t }) => {
+export default({ t, setting }) => {
   return [
     { keyName: 'sizeSmallCenter', type: BUTTONS.SIZE_SMALL_CENTER, mobile: false },
     { keyName: 'sizeContent', type: BUTTONS.SIZE_CONTENT, mobile: false },
