+++ packages/plugin-video/src/toolbar/inputModalStyles.js
@@ -1,0 +1,15 @@
+export const InputModalCustomStyle = {
+    content:
+    {
+      maxWidth: '580px',
+      minHeight: '280px'
+    }
+  };
+  
+export const ExtendedInputModalCustomStyle = {
+    content:
+    {
+      maxWidth: '580px',
+      minHeight: '378px'
+    }
+  };
