+++ packages/plugin-video/statics/styles/video-selection-input-modal.scss
@@ -1,0 +1,149 @@
+@import "~wix-rich-content-common/dist/statics/styles/mixins";
+@import "~wix-rich-content-common/dist/statics/styles/palette";
+
+.video_modal_container {
+  padding: 51px 60px 60px;
+  @include phone {
+    padding: 24px 18px 30px;
+  }
+}
+
+.video_modal_custom_container {
+  padding: 48px 60px 60px 60px;
+  @include phone {
+    padding: 24px 18px 30px;
+  }
+}
+
+.video_modal_closeIcon {
+  position: absolute;
+  right: 30px;
+  top: 30px;
+  cursor: pointer;
+}
+
+.video_modal_cameraIcon {
+  width: 18px;
+}
+
+.video_modal_errorIcon {
+  width: 13px;
+}
+
+.video_modal_header {
+  display: flex;
+  align-items: center;
+}
+
+.video_modal_header_text {
+  font-size: 16px;
+  margin: 8px 0 8px 0;
+}
+
+.video_modal_custom_header_text {
+  font-size: 16px;
+  height: 24px;
+  margin: 0px 0 12px 0;
+  width: 460px;
+  text-align: center;
+  @include phone {
+    margin: 0px 0 30px 0px;
+  }
+}
+
+.video_modal_header_icon {
+  margin-right: 12px;
+}
+
+.video_modal_textInput {
+  margin: 8px 0 24px 0;
+  input[type="url"] {
+    padding-right: 18px;
+  }
+}
+
+.video_modal_custom_textInput {
+  margin: 0px 0 48px 0px;
+  width: calc(100% - 119px);
+  display: inline-block;
+  input[type="url"] {
+    padding-right: 18px;
+  }
+  @include phone {
+    width: 100%;
+    margin: 0px 0 15px 0px;
+  }
+}
+
+.video_modal_footer {
+  border: none !important;
+  padding: 0 !important;
+  height: 40px !important;
+}
+
+.video_modal_add_a_Video {
+  width: 100%;
+  height: auto;
+  font-family: "Courier New", Courier, monospace;
+  font-size: 24px;
+  font-weight: 500;
+  font-style: normal;
+  font-stretch: normal;
+  letter-spacing: normal;
+  text-align: center;
+  color: #333333;
+  padding-bottom: 36px;
+  @include phone {
+    width: 100%;
+    padding-bottom: 20px;
+  }
+}
+
+.video_modal_or_upload_video_fro {
+  width: 100%;
+  height: 24px;
+  font-family: Avenir;
+  font-size: 16px;
+  font-weight: 500;
+  font-style: normal;
+  font-stretch: normal;
+  line-height: 1.5;
+  letter-spacing: normal;
+  text-align: center;
+  color: #333333;
+  padding-bottom: 24px;
+}
+
+.video_modal_upload_video {
+  width: 100%;
+  height: 24px;
+  font-family: Avenir;
+  font-size: 16px;
+  font-weight: 500;
+  font-style: normal;
+  font-stretch: normal;
+  line-height: 1.5;
+  letter-spacing: normal;
+  text-align: center;
+  color: #236fff;
+  display: inline-block;
+  cursor: pointer;
+}
+
+.video_modal_add_button {
+  width: 108px;
+  display: inline-block;
+  border: none;
+  padding-left: 24px;
+  padding-right: 24px;
+  margin-left: 10px;
+  height: 40px;
+  @include phone {
+    width: 40%; 
+    margin-left: 30%;
+    margin-bottom: 15px;
+  }
+}
+.video_modal_add_button:disabled {
+  opacity: 0.5;
+}
