--- packages/plugin-video/statics/styles/video-url-input-modal.scss
@@ -1,51 +1,0 @@
-@import '~wix-rich-content-common/dist/statics/styles/mixins';
-@import '~wix-rich-content-common/dist/statics/styles/palette';
-
-.container {
-    padding: 51px 60px 60px;
-    @include phone {
-        padding: 24px 18px 30px;
-    }
-}
-
-.closeIcon {
-    position: absolute;
-    right: 30px;
-    top: 30px;
-    cursor: pointer;
-}
-
-.cameraIcon {
-    width: 18px;
-}
-
-.errorIcon {
-    width: 13px;
-}
-
-.header {
-    display: flex;
-    align-items: center;
-}
-
-.header_text {
-    font-size: 16px;
-    margin: 8px 0 8px 0;
-}
-
-.header_icon {
-    margin-right: 12px;
-}
-
-.videoUrlInputModal_textInput {
-    margin: 8px 0 24px 0;
-    input[type = 'url'] {
-        padding-right: 18px;
-    } 
-}
-
-.modal_footer {
-    border: none !important;
-    padding: 0 !important;
-    height: 40px !important;
-}
